requirements.txt
# 核心依赖
python>=3.8
pyyaml>=6.0

# 数据库依赖
chromadb>=0.4.0

# 压缩依赖
# 可选：如果需要使用LZ4或Zstandard压缩
# lz4>=4.3.2
# zstandard>=0.22.0

# 开发依赖
pytest>=7.0.0
flake8>=6.0.0
black>=23.0.0